+++ Specs/DataSources/StaticGroundGeometryPerMaterialBatchSpec.js
@@ -338,4 +338,81 @@ defineSuite([
             batch.removeAllPrimitives();
         });
     });
+
+    it('has correct show attribute after rebuilding primitive', function() {
+        if (!GroundPrimitive.isSupported(scene)) {
+            return;
+        }
+
+        if (!GroundPrimitive.isSupported(scene) || !GroundPrimitive.supportsMaterials(scene)) {
+            // Don't fail if materials on GroundPrimitive not supported
+            return;
+        }
+
+        var batch = new StaticGroundGeometryPerMaterialBatch(scene.primitives, MaterialAppearance);
+
+        function buildEntity(x, y, z) {
+            var material = new GridMaterialProperty({
+                color : Color.YELLOW,
+                cellAlpha : 0.3,
+                lineCount : new Cartesian2(8, 8),
+                lineThickness : new Cartesian2(2.0, 2.0)
+            });
+
+            return new Entity({
+                position : new Cartesian3(x, y, z),
+                ellipse : {
+                    semiMajorAxis : 2,
+                    semiMinorAxis : 1,
+                    material: material
+                }
+            });
+        }
+
+        function renderScene() {
+            scene.initializeFrame();
+            var isUpdated = batch.update(time);
+            scene.render(time);
+            return isUpdated;
+        }
+
+        var entity1 = buildEntity(1234, 5678, 9101112);
+        var updater1 = new EllipseGeometryUpdater(entity1, scene);
+        batch.add(time, updater1);
+
+        var entity2 = buildEntity(123, 456, 789);
+        var updater2 = new EllipseGeometryUpdater(entity2, scene);
+
+        return pollToPromise(renderScene)
+            .then(function() {
+                expect(scene.primitives.length).toEqual(1);
+                var primitive = scene.primitives.get(0);
+                var attributes = primitive.getGeometryInstanceAttributes(entity1);
+                expect(attributes.show).toEqual([1]);
+
+                entity1.show = false;
+                updater1._onEntityPropertyChanged(entity1, 'isShowing');
+                return pollToPromise(renderScene);
+            })
+            .then(function() {
+                expect(scene.primitives.length).toEqual(1);
+                var primitive = scene.primitives.get(0);
+                var attributes = primitive.getGeometryInstanceAttributes(entity1);
+                expect(attributes.show).toEqual([0]);
+
+                batch.add(time, updater2);
+                return pollToPromise(renderScene);
+            })
+            .then(function() {
+                expect(scene.primitives.length).toEqual(1);
+                var primitive = scene.primitives.get(0);
+                var attributes = primitive.getGeometryInstanceAttributes(entity1);
+                expect(attributes.show).toEqual([0]);
+
+                attributes = primitive.getGeometryInstanceAttributes(entity2);
+                expect(attributes.show).toEqual([1]);
+
+                batch.removeAllPrimitives();
+            });
+    });
 });